+++ tests/test_signal_processing.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-from SignalProcessing.new import SignalProcessing, IntegrationRules, Windows
+from SignalProcessing.signal import SignalProcessing, IntegrationRules, Windows
 
 TOL = 3e-3
 FREQ = 6
